requirements.txt
hydra-core==0.11.3
loguru==0.5.0
numpy==1.18.4
omegaconf==1.4.1
pandas==0.25.3
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
tensorflow-gpu==1.15.0
tqdm==4.46.0
prettytable==0.7.2
